requirements.txt
streamlit
semantic-kernel
azure.identity
azure.ai.projects